Make Church Center event fetching more robust
Previously, if an event was visible on Church Center, it would be displayed, and start and end times would be fetched for it. When an event had no public times (say, for an all-day event), this would result in a server error. Now, the event fetching code simply ignores such events.