More realistic form conditions, first steps
This commit represents the first steps of what will likely be a good first implementation of conditionally shown fields. It won't require completely rebuilding the form logic, and should be fairly easy to get working. However, there's not time right now. Time to shelve it and get basic forms published.